+++ CHANGELOG.md
@@ -15,8 +15,8 @@
 
 <!-- Your comment below this -->
 
-- Added a CLI option `--ignoreInvalidInlineComments` so that you can ignore inline-comments for invalid lines (e.g
-  unchanged lines). The comment would be ignored completely - it won't even show in the results commit. [@pinkasey]
+- Added a CLI option `--noOutOfDiffComments` so that you can ignore inline-comments for lines that were not changed in
+  this PR. The comments would be ignored completely - they won't even show in the results comment. [@pinkasey]
 
 <!-- Your comment above this -->